+++ src/components/sidebar/pointOfSale/productCard.tsx
