+++ src/pages/dashboard/ayuda/index.tsx
@@ -1,0 +1,13 @@
+import DashboardLayout from "../../../layout";
+
+const AyudaPage = () => {
+    return (
+        <div>
+            Ayuda
+        </div>
+    );
+}
+
+AyudaPage.PageLayout = DashboardLayout;
+
+export default AyudaPage;
