+++ src/pages/dashboard/configuracion/index.tsx
@@ -1,0 +1,13 @@
+import DashboardLayout from "../../../layout";
+
+const ConfiguracionPage = () => {
+    return (
+        <div>
+            Ayuda
+        </div>
+    );
+}
+
+ConfiguracionPage.PageLayout = DashboardLayout;
+
+export default ConfiguracionPage;
